Revert "Merge pull request #52 from duncathan/feature/rework-script-api"
This reverts commit 96b761ecf657ce5236a7ee90eae17af0071fb4be.
