+++ retro_data_structures/formats/mrea.py
@@ -1,11 +1,8 @@
 """
 Wiki: https://wiki.axiodl.com/w/MREA_(Metroid_Prime_2)
 """
-from __future__ import annotations
-
 import copy
 import io
-import itertools
 import typing
 from collections.abc import Iterator
 from enum import IntEnum
@@ -25,30 +22,19 @@ from construct.lib.containers import Container, ListContainer
 
 from retro_data_structures import game_check
 from retro_data_structures.base_resource import AssetId, AssetType, BaseResource, Dependency
-from retro_data_structures.common_types import AssetId32, FourCC, Transform4f
+from retro_data_structures.common_types import FourCC, Transform4f
 from retro_data_structures.compression import LZOCompressedBlock
 from retro_data_structures.construct_extensions.alignment import PrefixedWithPaddingBefore
 from retro_data_structures.construct_extensions.version import BeforeVersion, WithVersion
 from retro_data_structures.data_section import DataSection
-from retro_data_structures.exceptions import UnknownAssetId
 from retro_data_structures.formats.area_collision import AreaCollision
 from retro_data_structures.formats.arot import AROT
-from retro_data_structures.formats.cmdl import dependencies_for_material_set
 from retro_data_structures.formats.lights import Lights
-from retro_data_structures.formats.script_layer import SCGN, SCLY, ScriptLayer, new_layer
-from retro_data_structures.formats.script_object import (
-    InstanceIdRef,
-    InstanceRef,
-    ScriptInstance,
-    resolve_instance_id_ref,
-)
-from retro_data_structures.formats.strg import Strg
+from retro_data_structures.formats.script_layer import SCGN, SCLY, ScriptLayerHelper
+from retro_data_structures.formats.script_object import InstanceId, ScriptInstanceHelper
 from retro_data_structures.formats.visi import VISI
 from retro_data_structures.formats.world_geometry import lazy_world_geometry
 from retro_data_structures.game_check import AssetIdCorrect, Game
-
-if typing.TYPE_CHECKING:
-    from retro_data_structures.formats.mlvl import Mlvl
 
 
 class MREAVersion(IntEnum):
@@ -437,7 +423,7 @@ MREA = MREAConstruct()
 
 
 class Mrea(BaseResource):
-    _script_layer_helpers: dict[int, ScriptLayer] | None = None
+    _script_layer_helpers: dict[int, ScriptLayerHelper] | None = None
 
     @classmethod
     def resource_type(cls) -> AssetType:
@@ -489,25 +475,21 @@ class Mrea(BaseResource):
                 self._raw.sections.script_layers_section[i] = section._raw
         return super().build()
 
-    def _all_non_scgn_instances(self) -> Iterator[ScriptInstance]:
-        for layer in self.script_layers:
-            yield from layer.instances
-
     @property
-    def script_layers(self) -> Iterator[ScriptLayer]:
+    def script_layers(self) -> Iterator[ScriptLayerHelper]:
         self._ensure_decoded_section("script_layers_section", lazy_load=self.target_game != Game.PRIME)
 
         if self.target_game == Game.PRIME:
             section = self._raw.sections.script_layers_section[0]
             for i, layer in enumerate(section.layers):
-                yield ScriptLayer(layer, i, self.target_game)
+                yield ScriptLayerHelper(layer, i, self.target_game)
         else:
             if self._script_layer_helpers is None:
                 self._script_layer_helpers = {}
 
             for i, section in enumerate(self._raw.sections.script_layers_section):
                 if i not in self._script_layer_helpers:
-                    self._script_layer_helpers[i] = ScriptLayer(
+                    self._script_layer_helpers[i] = ScriptLayerHelper(
                         _CATEGORY_ENCODINGS["script_layers_section"].parse(
                             section, target_game=self.target_game
                         ),
@@ -517,318 +499,45 @@ class Mrea(BaseResource):
 
             yield from self._script_layer_helpers.values()
 
-    _generated_objects_layer: ScriptLayer | None = None
+    _generated_objects_layer: ScriptLayerHelper | None = None
     @property
-    def generated_objects_layer(self) -> ScriptLayer:
+    def generated_objects_layer(self) -> ScriptLayerHelper:
         assert self.target_game >= Game.ECHOES
         if self._generated_objects_layer is None:
-            self._generated_objects_layer = ScriptLayer(
+            self._generated_objects_layer = ScriptLayerHelper(
                 self.get_section("generated_script_objects_section")[0],
                 None,
                 self.target_game
             )
         return self._generated_objects_layer
 
-
-_hardcoded_dependencies: dict[int, dict[str, list[Dependency]]] = {
-    0xD7C3B839: {
-        # Sanctum
-        "Default": [Dependency("TXTR", 0xd5b9e5d1)],
-        "Emperor Ing Stage 1": [Dependency("TXTR", 0x52c7d438)],
-        "Emperor Ing Stage 3": [Dependency("TXTR", 0xd5b9e5d1)],
-        "Emperor Ing Stage 1 Intro Cine": [Dependency("TXTR", 0x52c7d438)],
-        "Emperor Ing Stage 3 Death Cine": [Dependency("TXTR", 0xd5b9e5d1)],
-    },
-    0xA92F00B3: {
-        # Hive Temple
-        "CliffsideBoss": [
-            Dependency("TXTR", 0x24149e16),
-            Dependency("TXTR", 0xbdb8a88a),
-            Dependency("FSM2", 0x3d31822b),
-        ]
-    },
-    0xC0113CE8: {
-        # Dynamo Works
-        "3rd Pass": [Dependency("RULE", 0x393ca543)]
-    },
-    0x5571E89E: {
-        # Hall of Combat Mastery
-        "2nd Pass Enemies": [Dependency("RULE", 0x393ca543)]
-    },
-    0x7B94B06B: {
-        # Hive Portal Chamber
-        "1st Pass": [Dependency("RULE", 0x393ca543)],
-        "2nd Pass": [Dependency("RULE", 0x393ca543)]
-    },
-    0xF8DBC03D: {
-        # Hive Reactor
-        "2nd Pass": [Dependency("RULE", 0x393ca543)]
-    },
-    0xB666B655: {
-        # Reactor Access
-        "2nd Pass": [Dependency("RULE", 0x393ca543)]
-    },
-    0xE79AAFAE: {
-        # Transport A Access
-        "2nd Pass": [Dependency("RULE", 0x393ca543)]
-    },
-    0xFEB7BD27: {
-        # Transport B Access
-        "Default": [Dependency("RULE", 0x393ca543)]
-    },
-    0x89D246FD: {
-        # Portal Access
-        "Default": [Dependency("RULE", 0x393ca543)]
-    },
-    0x0253782D: {
-        # Dark Forgotten Bridge
-        "Default": [Dependency("RULE", 0x393ca543)]
-    },
-    0x09DECF21: {
-        # Forgotten Bridge
-        "Default": [Dependency("RULE", 0x393ca543)]
-    },
-    0x629790F4: {
-        # Sacrificial Chamber
-        "1st Pass": [Dependency("RULE", 0x393ca543)]
-    },
-    0xBBE4B3AE: {
-        # Dungeon
-        "Default": [Dependency("TXTR", 0xe252e7f6)]
-    },
-    0x2BCD44A7: {
-        # Portal Terminal
-        "Default": [Dependency("TXTR", 0xb6fa5023)]
-    },
-    0xC68B5B51: {
-        # Transport to Sanctuary Fortress
-        "!!non_layer!!": [Dependency("TXTR", 0x75a219a8)]
-    },
-    0x625A2692: {
-        # Temple Transport Access
-        "!!non_layer!!": [Dependency("TXTR", 0x581c56ea)]
-    },
-    0x96F4CA1E: {
-        # Minigyro Chamber
-        "Default": [Dependency("TXTR", 0xac080dfb)]
-    },
-    0x5BBF334F: {
-        # Staging Area
-        "!!non_layer!!": [Dependency("TXTR", 0x738feb19)]
-    }
-}
-
-
-class Area:
-    _mrea: Mrea = None
-    _strg: Strg = None
-
-    def __init__(self, raw: Container, index: int, parent_mlvl: Mlvl):
-        self._raw = raw
-        self.asset_manager = parent_mlvl.asset_manager
-        self._flags = parent_mlvl.layer_flags_for_area_index(index)
-        self._layer_names = parent_mlvl.layer_names_for_area_index(index)
-        self._index = index
-        self._parent_mlvl = parent_mlvl
-
     @property
-    def id(self) -> int:
-        return self._raw.internal_area_id
-
-    @property
-    def index(self) -> int:
-        return self._index
-
-    @property
-    def name(self) -> str:
-        try:
-            return self.strg.strings[0]
-        except UnknownAssetId:
-            return "!!" + self.internal_name
-
-    @name.setter
-    def name(self, value):
-        self.strg.set_string(0, value)
-
-    @property
-    def internal_name(self) -> str:
-        return self._raw.get("internal_area_name", "Unknown")
-
-    @property
-    def strg(self) -> Strg:
-        if self._strg is None:
-            self._strg = self.asset_manager.get_file(self._raw.area_name_id, type_hint=Strg)
-        return self._strg
-
-    @property
-    def mrea(self) -> Mrea:
-        if self._mrea is None:
-            self._mrea = self.asset_manager.get_file(self.mrea_asset_id, type_hint=Mrea)
-        return self._mrea
-
-    @property
-    def mrea_asset_id(self) -> int:
-        return self._raw.area_mrea_id
-
-    @property
-    def layers(self) -> Iterator[ScriptLayer]:
-        for layer in self.mrea.script_layers:
-            yield layer.with_parent(self)
-
-    @property
-    def generated_objects_layer(self) -> ScriptLayer:
-        return self.mrea.generated_objects_layer.with_parent(self)
-
-    def get_layer(self, name: str) -> ScriptLayer:
-        return next(layer for layer in self.layers if layer.name == name)
-
-    def add_layer(self, name: str, active: bool = True) -> ScriptLayer:
-        index = len(self._layer_names)
-        self._layer_names.append(name)
-        self._flags.append(active)
-        raw = new_layer(index, self.asset_manager.target_game)
-        self.mrea._raw.sections.script_layer_section.append(raw)
-        return self.get_layer(name)
-
-    @property
-    def next_instance_id(self) -> int:
-        ids = [instance.id.instance for instance in self.all_instances]
-        return next(i for i in itertools.count() if i not in ids)
-
-    @property
-    def all_instances(self) -> Iterator[ScriptInstance]:
-        for layer in self.layers:
+    def all_instances(self) -> Iterator[ScriptInstanceHelper]:
+        for layer in self.script_layers:
             yield from layer.instances
-        yield from self.generated_objects_layer.instances
 
-    def get_instance(self, instance: InstanceRef) -> ScriptInstance:
-        if (inst := self.generated_objects_layer.get_instance(instance, must_exist=False)):
-            return inst
-        for layer in self.layers:
-            if (inst := layer.get_instance(instance, must_exist=False)) is not None:
-                return inst
+    def get_instance(self, instance_id: int) -> ScriptInstanceHelper | None:
+        for layer in self.script_layers:
+            if (instance := layer.get_instance(instance_id)) is not None:
+                return instance
+        if (instance := self.generated_objects_layer.get_instance(instance_id)) is not None:
+            return instance
 
-        if isinstance(instance, InstanceIdRef):
-            instance = resolve_instance_id_ref(instance)
-        raise KeyError(f"No instance {instance} found")
+    def get_instance_by_name(self, name: str) -> ScriptInstanceHelper:
+        for layer in self.script_layers:
+            if (instance := layer.get_instance_by_name(name, raise_if_missing=False)) is not None:
+                return instance
+        if (instance := self.generated_objects_layer.get_instance_by_name(name, raise_if_missing=False)) is not None:
+            return instance
+        raise KeyError(name)
 
-    def get_layer_for_instance(self, instance: InstanceRef) -> ScriptLayer:
-        if self.generated_objects_layer.has_instance(instance):
-            return self.generated_objects_layer
-        return next(layer for layer in self.layers if layer.has_instance(instance))
+    def remove_instance(self, instance: int | InstanceId | str | ScriptInstanceHelper):
+        if isinstance(instance, str):
+            instance = self.get_instance(instance)
+        elif isinstance(instance, int):
+            instance = InstanceId(instance)
+        if isinstance(instance, ScriptInstanceHelper):
+            instance = instance.id
 
-    def remove_instance(self, instance: InstanceRef):
-        self.get_layer_for_instance(instance).remove_instance(instance)
-
-    def _raw_connect_to(self, source_dock_number: int, target_area: Area, target_dock_number: int):
-        source_dock = self._raw.docks[source_dock_number]
-        assert len(source_dock.connecting_dock) == 1, "Only docks with one connection supported"
-        source_dock.connecting_dock[0].area_index = target_area._index
-        source_dock.connecting_dock[0].dock_index = target_dock_number
-
-        attached_area_index = []
-        for docks in self._raw.docks:
-            for c in docks.connecting_dock:
-                if c.area_index not in attached_area_index:
-                    attached_area_index.append(c.area_index)
-        self._raw.attached_area_index = construct.ListContainer(attached_area_index)
-
-    def connect_dock_to(self, source_dock_number: int, target_area: Area, target_dock_number: int):
-        self._raw_connect_to(source_dock_number, target_area, target_dock_number)
-        target_area._raw_connect_to(target_dock_number, self, source_dock_number)
-
-    def build_non_layer_dependencies(self) -> typing.Iterator[Dependency]:
-        if self.asset_manager.target_game <= Game.ECHOES:
-            geometry_section = self.mrea.get_raw_section("geometry_section")
-            if geometry_section:
-                for asset_id in PrefixedArray(Int32ub, AssetId32).parse(geometry_section[0]):
-                    yield from self.asset_manager.get_dependencies_for_asset(asset_id)
-        else:
-            geometry = self.mrea.get_geometry()
-            if geometry is not None:
-                yield from dependencies_for_material_set(geometry[0].materials, self.asset_manager)
-
-        valid_asset = self.asset_manager.target_game.is_valid_asset_id
-        if valid_asset(portal_area := self.mrea.get_portal_area()):
-            yield Dependency("PTLA", portal_area)
-        if valid_asset(static_geometry_map := self.mrea.get_static_geometry_map()):
-            yield Dependency("EGMC", static_geometry_map)
-        if valid_asset(path := self.mrea.get_path()):
-            yield Dependency("PATH", path)
-
-    def build_scgn_dependencies(self, layer_deps: list[list[Dependency]], only_modified: bool = False):
-        layer_deps = list(layer_deps)
-
-        layers = list(self.layers)
-        for instance in self.generated_objects_layer.instances:
-            inst_layer = instance.id.layer
-            if not only_modified or layers[inst_layer].is_modified:
-                layer_deps[inst_layer].extend(instance.mlvl_dependencies_for(self.asset_manager))
-
-        return [list(dict.fromkeys(deps)) for deps in layer_deps]
-
-    def build_mlvl_dependencies(self, only_modified: bool = False):
-        layer_deps = [
-            list(
-                layer.build_mlvl_dependencies(self.asset_manager)
-                if (not only_modified) or layer.is_modified() else
-                layer.dependencies
-            ) for layer in self.layers
-        ]
-
-        if only_modified:
-            # assume we never modify these
-            layer_deps.append(list(self.non_layer_dependencies))
-        else:
-            non_layer_deps = list(self.build_non_layer_dependencies())
-            if "!!non_layer!!" in _hardcoded_dependencies.get(self.mrea_asset_id, {}):
-                non_layer_deps.extend(_hardcoded_dependencies[self.mrea_asset_id]["!!non_layer!!"])
-            layer_deps.append(non_layer_deps)
-
-        layer_deps = self.build_scgn_dependencies(layer_deps, only_modified)
-
-        if self.mrea_asset_id in _hardcoded_dependencies:
-            for layer_name, missing in _hardcoded_dependencies[self.mrea_asset_id].items():
-                if layer_name == "!!non_layer!!":
-                    continue
-
-                layer = self.get_layer(layer_name)
-                if only_modified and not layer.is_modified:
-                    continue
-
-                layer_deps[layer.index].extend(missing)
-
-        layer_deps = [
-            [dep for dep in layer if not dep.exclude_for_mlvl]
-            for layer in layer_deps
-        ]
-
-        offset = 0
-        offsets = []
-        for layer_dep in layer_deps:
-            offsets.append(offset)
-            offset += len(layer_dep)
-
-        fancy_deps: list[Dependency] = list(itertools.chain(*layer_deps))
-        deps = [Container(asset_type=dep.type, asset_id=dep.id) for dep in fancy_deps]
-        self._raw.dependencies.dependencies = deps
-        self._raw.dependencies.offsets = offsets
-
-    @property
-    def layer_dependencies(self) -> dict[str, list[Dependency]]:
-        return {
-            layer.name: list(layer.dependencies)
-            for layer in self.layers
-        }
-
-    @property
-    def non_layer_dependencies(self) -> Iterator[Dependency]:
-        deps = self._raw.dependencies
-        global_deps = deps.dependencies[deps.offsets[len(self._layer_names)]:]
-        yield from [Dependency(dep.asset_type, dep.asset_id) for dep in global_deps]
-
-    @property
-    def dependencies(self) -> dict[str, list[Dependency]]:
-        deps = self.layer_dependencies
-        deps["!!non_layer!!"] = list(self.non_layer_dependencies)
-        return deps
+        layers = list(self.script_layers)
+        layers[instance.layer].remove_instance(instance)
