+++ retro_data_structures/formats/script_object.py
@@ -34,7 +34,7 @@ from retro_data_structures.game_check import Game
 
 if TYPE_CHECKING:
     from retro_data_structures.asset_manager import AssetManager
-    from retro_data_structures.formats.script_layer import ScriptLayer
+    from retro_data_structures.formats.script_layer import ScriptLayerHelper
     from retro_data_structures.properties.base_property import BaseObjectType
 
     PropertyType = typing.TypeVar("PropertyType", bound=BaseObjectType)
@@ -214,7 +214,7 @@ def _try_quick_get_name(data: bytes) -> str | None:
         return None
 
 
-class ScriptInstance:
+class ScriptInstanceHelper:
     _raw: ScriptInstanceRaw
     target_game: Game
 
@@ -227,10 +227,10 @@ class ScriptInstance:
         return f"<ScriptInstance {self.type_name} 0x{self.id:08x}>"
 
     def __eq__(self, other):
-        return isinstance(other, ScriptInstance) and self._raw == other._raw
+        return isinstance(other, ScriptInstanceHelper) and self._raw == other._raw
 
     @classmethod
-    def new_instance(cls, target_game: Game, instance_type: str, layer: ScriptLayer) -> ScriptInstance:
+    def new_instance(cls, target_game: Game, instance_type: str, layer: ScriptLayerHelper) -> ScriptInstanceHelper:
         property_type = properties.get_game_object(target_game, instance_type)
 
         raw = ScriptInstanceRaw(
@@ -242,7 +242,7 @@ class ScriptInstance:
         return cls(raw, target_game, on_modify=layer.mark_modified)
 
     @classmethod
-    def new_from_properties(cls, object_properties: BaseObjectType, layer: ScriptLayer) -> ScriptInstance:
+    def new_from_properties(cls, object_properties: BaseObjectType, layer: ScriptLayerHelper) -> ScriptInstanceHelper:
         raw = ScriptInstanceRaw(
             type=object_properties.object_type(),
             id=layer.new_instance_id(),
@@ -268,9 +268,11 @@ class ScriptInstance:
         self._raw.id = InstanceId(value)
         self.on_modify()
 
-    def id_matches(self, other: InstanceIdRef) -> bool:
-        other = resolve_instance_id_ref(other)
-        return self.id.area == other.area and self.id.instance == other.instance
+    def id_matches(self, id: typing.Union[int, InstanceId]) -> bool:
+        if not isinstance(id, InstanceId):
+            id = InstanceId(id)
+
+        return self.id.area == id.area and self.id.instance == id.instance
 
     @property
     def name(self) -> str | None:
@@ -330,7 +332,7 @@ class ScriptInstance:
         self._raw.connections = tuple(value)
         self.on_modify()
 
-    def add_connection(self, state: str | State, message: str | Message, target: ScriptInstance):
+    def add_connection(self, state: str | State, message: str | Message, target: ScriptInstanceHelper):
         correct_state = enum_helper.STATE_PER_GAME[self.target_game]
         correct_message = enum_helper.MESSAGE_PER_GAME[self.target_game]
 
@@ -343,24 +345,11 @@ class ScriptInstance:
     def remove_connection(self, connection: Connection):
         self.connections = [c for c in self.connections if c is not connection]
 
-    def remove_connections(self, target: Union[int, ScriptInstance]):
-        if isinstance(target, ScriptInstance):
+    def remove_connections(self, target: Union[int, ScriptInstanceHelper]):
+        if isinstance(target, ScriptInstanceHelper):
             target = target.id
 
         self.connections = [c for c in self.connections if c.target != target]
 
     def mlvl_dependencies_for(self, asset_manager: AssetManager) -> Iterator[Dependency]:
         yield from self.get_properties().dependencies_for(asset_manager)
-
-
-InstanceIdRef = int | InstanceId | ScriptInstance
-InstanceRef = InstanceIdRef | str
-
-def resolve_instance_id_ref(inst: InstanceIdRef) -> InstanceId:
-    if isinstance(inst, InstanceId):
-        return inst
-    if isinstance(inst, ScriptInstance):
-        return inst.id
-    if isinstance(inst, int):
-        return InstanceId(inst)
-    raise TypeError(f"Invalid type: Expected InstanceIdRef, got {type(inst)}")
